Refactored into smaller classes
